requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
discord.py==1.3.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
librus-tricks==0.8.1
mccabe==0.6.1
multidict==4.7.5
pycparser==2.20
pylint==2.4.4
PyNaCl==1.3.0
requests==2.23.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
typed-ast==1.4.1
urllib3==1.25.9
websockets==8.1
wrapt==1.11.2
yarl==1.4.2